api/requirements.txt
alembic==1.5.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
Authlib==0.15.3
autopep8==1.5.4
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
ecdsa==0.14.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
iniconfig==1.1.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
